Add tests for CreateAuthnRequest
