+++ saml/models/core/response.go
@@ -16,7 +16,14 @@ type Response struct {
 
 // See 3.2.2 http://docs.oasis-open.org/security/saml/v2.0/saml-core-2.0-os.pdf
 type StatusResponseType struct {
-	RequestResponseCommon
+	ID           string      `xml:",attr"`           // required
+	Version      string      `xml:",attr"`           // required
+	IssueInstant time.Time   `xml:",attr"`           // required
+	Consent      string      `xml:",attr,omitempty"` // optional TODO: define constants
+	Issuer       *Issuer     // recommended
+	Singature    string      // recommended
+	Extensions   *Extensions // optional
+	Destination  string      `xml:",attr"`
 }
 
 // See 3.2.2.1 http://docs.oasis-open.org/security/saml/v2.0/saml-core-2.0-os.pdf
