+++ saml/options.go
@@ -1,0 +1,16 @@
+package saml
+
+// Option defines a common functional options type which can be used in a
+// variadic parameter pattern.
+type Option func(interface{})
+
+// ApplyOpts takes a pointer to the options struct as a set of default options
+// and applies the slice of opts as overrides.
+func ApplyOpts(opts interface{}, opt ...Option) {
+	for _, o := range opt {
+		if o == nil { // ignore any nil Options
+			continue
+		}
+		o(opts)
+	}
+}
